requirements.txt
fastapi==0.112.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-dotenv==1.*
httpx==0.27.*
postgrest==0.16.*
supabase==2.*
pandas==2.*
numpy==2.*
